requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
spacy>=3.7.2
pdfplumber>=0.10.3
python-multipart>=0.0.9
python-dotenv>=1.0.0
aiosmtplib>=3.0.1
jinja2>=3.1.3
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.3
